[Fix] force cast to integer per get_user_meta('lm-follower-counter') e get_user_meta('lm-following-counter')

+++ src/Utility/LMWPFollowerChaceCounter.php
@@ -15,13 +15,13 @@ trait LMWPFollowerChaceCounter
     // con l'informazione eventuale del numero di like associata
     private function incrementFollowerCounters($followerId, $followingId, $keyFollower, $keyFollowing)
     {
-        $count = get_user_meta($followerId, $keyFollowing, true);
+        $count = (int) get_user_meta($followerId, $keyFollowing, true);
         if (!is_numeric($count)) {
             update_user_meta($followerId, $keyFollowing, 1);
         }
         update_user_meta($followerId, $keyFollowing, ($count + 1));
 
-        $count = get_user_meta($followingId, $keyFollower, true);
+        $count = (int) get_user_meta($followingId, $keyFollower, true);
         if (!is_numeric($count)) {
             update_user_meta($followingId, $keyFollower, 1);
         }
@@ -32,13 +32,13 @@ trait LMWPFollowerChaceCounter
 
     private function decrementFollowerCounters($followerId, $followingId, $keyFollower, $keyFollowing)
     {
-        $count = get_user_meta($followerId, $keyFollowing, true);
+        $count = (int) get_user_meta($followerId, $keyFollowing, true);
         if (!is_numeric($count) || $count <= 1) {
             update_user_meta($followerId, $keyFollowing, 0);
         }
         update_user_meta($followerId, $keyFollowing, ($count - 1));
 
-        $count = get_user_meta($followingId, $keyFollower, true);
+        $count = (int) get_user_meta($followingId, $keyFollower, true);
         if (!is_numeric($count) || $count <= 1) {
             update_user_meta($followingId, $keyFollower, 0);
         }
